NetworkServiceProtocol and NetworkError-enum are added to the Network Service module, for proper functionality
